cross entropy loss function bugs fixed

+++ Optimization/Loss.py
@@ -8,11 +8,11 @@ class CrossEntropyLoss:
 
     def forward(self, prediction_tensor, label_tensor):
         self.prediction_tensor = prediction_tensor
-        loss = label_tensor * -np.log(prediction_tensor + np.finfo(np.float).eps)
+        loss = label_tensor * -np.log(prediction_tensor + np.finfo(float).eps)
         loss = np.sum(loss, axis=0, keepdims=False)
         loss = np.sum(loss, axis=0, keepdims=False)
         return loss
 
     def backward(self, label_tensor):
-        error_tensor = (-label_tensor) / (self.prediction_tensor + np.finfo(np.float).eps)
+        error_tensor = (-label_tensor) / (self.prediction_tensor + np.finfo(float).eps)
         return error_tensor
